+++ src/QueryParameters/PaginationParams.php
@@ -7,7 +7,7 @@ namespace Storyblok\ManagementApi\QueryParameters;
 class PaginationParams
 {
     public function __construct(
-        private readonly int $page = 1,
+        private int $page = 1,
         private readonly int $perPage = 25,
     ) {}
 
@@ -19,4 +19,21 @@ class PaginationParams
     {
         return ['page' => $this->page, 'per_page' => $this->perPage];
     }
+
+    public function page(): int
+    {
+        return $this->page;
+    }
+
+    public function perPage(): int
+    {
+        return $this->perPage;
+    }
+
+    public function incrementPage(): void
+    {
+        ++$this->page;
+    }
+
+
 }